refactored/combined submitSpotifyUserMe into setCurrentUser in userActions
