+++ client/src/app/reducers/spotifyApi.js
@@ -1,5 +1,5 @@
 import { SPOTIFY_WEB_API } from "../actions/spotifyApiActions";
-import { SPOTIFY_USER_ME } from "../actions/spotifyApiActions";
+// import { SPOTIFY_USER_ME } from "../actions/spotifyApiActions";
 
 const initialState = {
   accessToken: "",
@@ -15,11 +15,11 @@ export const spotifyWebApi = (state = initialState, action) => {
         accessToken: action.payload.access_token,
         refreshToken: action.payload.refresh_token,
       };
-    case SPOTIFY_USER_ME:
-      return {
-        ...state,
-        user: action.payload,
-      };
+    // case SPOTIFY_USER_ME:
+    //   return {
+    //     ...state,
+    //     user: action.payload,
+    //   };
     default:
       return state;
   }
